introduced Tag interface
