+++ core/src/jsMain/kotlin/dev/fritz2/dom/html/elements.kt
@@ -2,20 +2,11 @@
 
 package dev.fritz2.dom.html
 
-import dev.fritz2.binding.Patch
-import dev.fritz2.binding.Store
 import dev.fritz2.binding.mountSimple
-import dev.fritz2.binding.sub
-import dev.fritz2.dom.*
-import dev.fritz2.lenses.IdProvider
-import dev.fritz2.utils.Myer
+import dev.fritz2.dom.HtmlTag
 import kotlinx.browser.document
 import kotlinx.coroutines.Job
-import kotlinx.coroutines.cancelChildren
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.scan
-import kotlinx.dom.clear
 import org.w3c.dom.*
 import org.w3c.dom.svg.SVGElement
 import org.w3c.dom.svg.SVGPathElement
@@ -24,8 +15,7 @@ import org.w3c.dom.svg.SVGPathElement
  * Exposes the JavaScript [HTMLAnchorElement](https://developer.mozilla.org/en/docs/Web/API/HTMLAnchorElement) to Kotlin
  */
 open class A(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLAnchorElement>("a", id, baseClass, job, scope),
-    WithText<HTMLAnchorElement> {
+    HtmlTag<HTMLAnchorElement>("a", id, baseClass, job, scope) {
     fun target(value: String) = attr("target", value)
     fun target(value: Flow<String>) = attr("target", value)
 
@@ -84,8 +74,7 @@ open class A(id: String? = null, baseClass: String? = null, job: Job, scope: Sco
  * Exposes the JavaScript [HTMLAreaElement](https://developer.mozilla.org/en/docs/Web/API/HTMLAreaElement) to Kotlin
  */
 open class Area(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLAreaElement>("area", id, baseClass, job, scope),
-    WithText<HTMLAreaElement> {
+    HtmlTag<HTMLAreaElement>("area", id, baseClass, job, scope) {
     fun alt(value: String) = attr("alt", value)
     fun alt(value: Flow<String>) = attr("alt", value)
 
@@ -147,15 +136,14 @@ open class Area(id: String? = null, baseClass: String? = null, job: Job, scope: 
  * Exposes the JavaScript [HTMLBRElement](https://developer.mozilla.org/en/docs/Web/API/HTMLBRElement) to Kotlin
  */
 open class Br(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLBRElement>("br", id, baseClass, job, scope)
+    HtmlTag<HTMLBRElement>("br", id, baseClass, job, scope)
 
 
 /**
  * Exposes the JavaScript [HTMLButtonElement](https://developer.mozilla.org/en/docs/Web/API/HTMLButtonElement) to Kotlin
  */
 open class Button(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLButtonElement>("button", id, baseClass, job, scope),
-    WithText<HTMLButtonElement> {
+    HtmlTag<HTMLButtonElement>("button", id, baseClass, job, scope) {
     fun autofocus(value: Boolean, trueValue: String = "") = attr("autofocus", value, trueValue)
     fun autofocus(value: Flow<Boolean>, trueValue: String = "") = attr("autofocus", value, trueValue)
 
@@ -192,8 +180,7 @@ open class Button(id: String? = null, baseClass: String? = null, job: Job, scope
  * Exposes the JavaScript [HTMLCanvasElement](https://developer.mozilla.org/en/docs/Web/API/HTMLCanvasElement) to Kotlin
  */
 open class Canvas(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLCanvasElement>("canvas", id, baseClass, job, scope),
-    WithText<HTMLCanvasElement> {
+    HtmlTag<HTMLCanvasElement>("canvas", id, baseClass, job, scope) {
     fun width(value: Int) = attr("width", value)
     fun width(value: Flow<Int>) = attr("width", value)
 
@@ -207,16 +194,14 @@ open class Canvas(id: String? = null, baseClass: String? = null, job: Job, scope
  * Exposes the JavaScript [HTMLDListElement](https://developer.mozilla.org/en/docs/Web/API/HTMLDListElement) to Kotlin
  */
 open class Dl(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLDListElement>("dl", id, baseClass, job, scope),
-    WithText<HTMLDListElement>
+    HtmlTag<HTMLDListElement>("dl", id, baseClass, job, scope)
 
 
 /**
  * Exposes the JavaScript [HTMLDataElement](https://developer.mozilla.org/en/docs/Web/API/HTMLDataElement) to Kotlin
  */
 open class Data(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLDataElement>("data", id, baseClass, job, scope),
-    WithText<HTMLDataElement> {
+    HtmlTag<HTMLDataElement>("data", id, baseClass, job, scope) {
     fun value(value: String) = attr("value", value)
     fun value(value: Flow<String>) = attr("value", value)
 }
@@ -226,16 +211,14 @@ open class Data(id: String? = null, baseClass: String? = null, job: Job, scope: 
  * Exposes the JavaScript [HTMLDataListElement](https://developer.mozilla.org/en/docs/Web/API/HTMLDataListElement) to Kotlin
  */
 open class DataList(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLDataListElement>("datalist", id, baseClass, job, scope),
-    WithText<HTMLDataListElement>
+    HtmlTag<HTMLDataListElement>("datalist", id, baseClass, job, scope) 
 
 
 /**
  * Exposes the JavaScript [HTMLDetailsElement](https://developer.mozilla.org/en/docs/Web/API/HTMLDetailsElement) to Kotlin
  */
 open class Details(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLDetailsElement>("details", id, baseClass, job, scope),
-    WithText<HTMLDetailsElement> {
+    HtmlTag<HTMLDetailsElement>("details", id, baseClass, job, scope) {
     fun open(value: Boolean, trueValue: String = "") = attr("open", value, trueValue)
     fun open(value: Flow<Boolean>, trueValue: String = "") = attr("open", value, trueValue)
 
@@ -246,8 +229,7 @@ open class Details(id: String? = null, baseClass: String? = null, job: Job, scop
  * Exposes the JavaScript [HTMLDialogElement](https://developer.mozilla.org/en/docs/Web/API/HTMLDialogElement) to Kotlin
  */
 open class Dialog(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLDialogElement>("dialog", id, baseClass, job, scope),
-    WithText<HTMLDialogElement> {
+    HtmlTag<HTMLDialogElement>("dialog", id, baseClass, job, scope) {
     fun open(value: Boolean, trueValue: String = "") = attr("open", value, trueValue)
     fun open(value: Flow<Boolean>, trueValue: String = "") = attr("open", value, trueValue)
 
@@ -261,8 +243,7 @@ open class Dialog(id: String? = null, baseClass: String? = null, job: Job, scope
  * Exposes the JavaScript [HTMLDivElement](https://developer.mozilla.org/en/docs/Web/API/HTMLDivElement) to Kotlin
  */
 open class Div(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLDivElement>("div", id, baseClass, job, scope),
-    WithText<HTMLDivElement> {
+    HtmlTag<HTMLDivElement>("div", id, baseClass, job, scope) {
     fun align(value: String) = attr("align", value)
     fun align(value: Flow<String>) = attr("align", value)
 
@@ -273,7 +254,7 @@ open class Div(id: String? = null, baseClass: String? = null, job: Job, scope: S
  * Exposes the JavaScript [HTMLDivElement](https://developer.mozilla.org/en/docs/Web/API/HTMLDivElement) to Kotlin
  */
 open class Embed(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLEmbedElement>("embed", id, baseClass, job, scope) {
+    HtmlTag<HTMLEmbedElement>("embed", id, baseClass, job, scope) {
     fun src(value: String) = attr("src", value)
     fun src(value: Flow<String>) = attr("src", value)
 
@@ -293,7 +274,7 @@ open class Embed(id: String? = null, baseClass: String? = null, job: Job, scope:
  * Exposes the JavaScript [HTMLFieldSetElement](https://developer.mozilla.org/en/docs/Web/API/HTMLFieldSetElement) to Kotlin
  */
 open class FieldSet(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLFieldSetElement>("fieldSet", id, baseClass, job, scope) {
+    HtmlTag<HTMLFieldSetElement>("fieldSet", id, baseClass, job, scope) {
     fun disabled(value: Boolean, trueValue: String = "") = attr("disabled", value, trueValue)
     fun disabled(value: Flow<Boolean>, trueValue: String = "") = attr("disabled", value, trueValue)
 
@@ -307,8 +288,7 @@ open class FieldSet(id: String? = null, baseClass: String? = null, job: Job, sco
  * Exposes the JavaScript [HTMLFormElement](https://developer.mozilla.org/en/docs/Web/API/`for`mElement) to Kotlin
  */
 open class Form(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLFormElement>("form", id, baseClass, job, scope),
-    WithText<HTMLFormElement> {
+    HtmlTag<HTMLFormElement>("form", id, baseClass, job, scope) {
     fun acceptCharset(value: String) = attr("acceptCharset", value)
     fun acceptCharset(value: Flow<String>) = attr("acceptCharset", value)
 
@@ -343,23 +323,21 @@ open class Form(id: String? = null, baseClass: String? = null, job: Job, scope: 
  * Exposes the JavaScript [HTMLHRElement](https://developer.mozilla.org/en/docs/Web/API/HTMLHRElement) to Kotlin
  */
 open class Hr(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLHRElement>("hr", id, baseClass, job, scope)
+    HtmlTag<HTMLHRElement>("hr", id, baseClass, job, scope)
 
 
 /**
  * Exposes the JavaScript [HTMLHeadingElement](https://developer.mozilla.org/en/docs/Web/API/HTMLHeadingElement) to Kotlin
  */
 open class H(num: Int, id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLHeadingElement>("h$num", id, baseClass, job, scope),
-    WithText<HTMLHeadingElement>
+    HtmlTag<HTMLHeadingElement>("h$num", id, baseClass, job, scope)
 
 
 /**
  * Exposes the JavaScript [HTMLIFrameElement](https://developer.mozilla.org/en/docs/Web/API/HTMLIFrameElement) to Kotlin
  */
 open class IFrame(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLIFrameElement>("iframe", id, baseClass, job, scope),
-    WithText<HTMLIFrameElement> {
+    HtmlTag<HTMLIFrameElement>("iframe", id, baseClass, job, scope) {
     fun src(value: String) = attr("src", value)
     fun src(value: Flow<String>) = attr("src", value)
 
@@ -391,8 +369,7 @@ open class IFrame(id: String? = null, baseClass: String? = null, job: Job, scope
  * Exposes the JavaScript [HTMLImageElement](https://developer.mozilla.org/en/docs/Web/API/HTMLImageElement) to Kotlin
  */
 open class Img(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLImageElement>("img", id, baseClass, job, scope),
-    WithText<HTMLImageElement> {
+    HtmlTag<HTMLImageElement>("img", id, baseClass, job, scope) {
     fun alt(value: String) = attr("alt", value)
     fun alt(value: Flow<String>) = attr("alt", value)
 
@@ -430,7 +407,7 @@ open class Img(id: String? = null, baseClass: String? = null, job: Job, scope: S
  * Exposes the JavaScript [HTMLInputElement](https://developer.mozilla.org/en/docs/Web/API/HTMLInputElement) to Kotlin
  */
 open class Input(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLInputElement>("input", id, baseClass, job, scope) {
+    HtmlTag<HTMLInputElement>("input", id, baseClass, job, scope) {
     fun accept(value: String) = attr("accept", value)
     fun accept(value: Flow<String>) = attr("accept", value)
 
@@ -552,8 +529,7 @@ open class Input(id: String? = null, baseClass: String? = null, job: Job, scope:
  * Exposes the JavaScript [HTMLLIElement](https://developer.mozilla.org/en/docs/Web/API/HTMLLIElement) to Kotlin
  */
 open class Li(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLLIElement>("li", id, baseClass, job, scope),
-    WithText<HTMLLIElement> {
+    HtmlTag<HTMLLIElement>("li", id, baseClass, job, scope) {
     fun value(value: Int) = attr("value", value)
     fun value(value: Flow<Int>) = attr("value", value)
 
@@ -564,8 +540,7 @@ open class Li(id: String? = null, baseClass: String? = null, job: Job, scope: Sc
  * Exposes the JavaScript [HTMLLabelElement](https://developer.mozilla.org/en/docs/Web/API/HTMLLabelElement) to Kotlin
  */
 open class Label(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLLabelElement>("label", id, baseClass, job, scope),
-    WithText<HTMLLabelElement> {
+    HtmlTag<HTMLLabelElement>("label", id, baseClass, job, scope) {
     fun `for`(value: String) = attr("for", value)
     fun `for`(value: Flow<String>) = attr("for", value)
 
@@ -576,16 +551,14 @@ open class Label(id: String? = null, baseClass: String? = null, job: Job, scope:
  * Exposes the JavaScript [HTMLLegendElement](https://developer.mozilla.org/en/docs/Web/API/HTMLLegendElement) to Kotlin
  */
 open class Legend(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLLegendElement>("legend", id, baseClass, job, scope),
-    WithText<HTMLLegendElement>
+    HtmlTag<HTMLLegendElement>("legend", id, baseClass, job, scope)
 
 
 /**
  * Exposes the JavaScript [HTMLMapElement](https://developer.mozilla.org/en/docs/Web/API/HTMLMapElement) to Kotlin
  */
 open class Map(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLMapElement>("map", id, baseClass, job, scope),
-    WithText<HTMLMapElement> {
+    HtmlTag<HTMLMapElement>("map", id, baseClass, job, scope) {
     fun name(value: String) = attr("name", value)
     fun name(value: Flow<String>) = attr("name", value)
 
@@ -596,8 +569,7 @@ open class Map(id: String? = null, baseClass: String? = null, job: Job, scope: S
  * Exposes the JavaScript [HTMLAudioElement](https://developer.mozilla.org/en/docs/Web/API/HTMLAudioElement) to Kotlin
  */
 open class Audio(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLAudioElement>("audio", id, baseClass, job, scope),
-    WithText<HTMLAudioElement> {
+    HtmlTag<HTMLAudioElement>("audio", id, baseClass, job, scope) {
     fun src(value: String) = attr("src", value)
     fun src(value: Flow<String>) = attr("src", value)
 
@@ -652,8 +624,7 @@ open class Audio(id: String? = null, baseClass: String? = null, job: Job, scope:
  * Exposes the JavaScript [HTMLVideoElement](https://developer.mozilla.org/en/docs/Web/API/HTMLVideoElement) to Kotlin
  */
 open class Video(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLVideoElement>("video", id, baseClass, job, scope),
-    WithText<HTMLVideoElement> {
+    HtmlTag<HTMLVideoElement>("video", id, baseClass, job, scope) {
     fun width(value: Int) = attr("width", value)
     fun width(value: Flow<Int>) = attr("width", value)
 
@@ -720,8 +691,7 @@ open class Video(id: String? = null, baseClass: String? = null, job: Job, scope:
  * Exposes the JavaScript [HTMLMeterElement](https://developer.mozilla.org/en/docs/Web/API/HTMLMeterElement) to Kotlin
  */
 open class Meter(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLMeterElement>("meter", id, baseClass, job, scope),
-    WithText<HTMLMeterElement> {
+    HtmlTag<HTMLMeterElement>("meter", id, baseClass, job, scope) {
     fun value(value: Double) = attr("value", value)
     fun value(value: Flow<Double>) = attr("value", value)
 
@@ -747,8 +717,7 @@ open class Meter(id: String? = null, baseClass: String? = null, job: Job, scope:
  * Exposes the JavaScript [HTMLModElement](https://developer.mozilla.org/en/docs/Web/API/HTMLModElement) to Kotlin
  */
 open class Ins(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLModElement>("ins", id, baseClass, job, scope),
-    WithText<HTMLModElement> {
+    HtmlTag<HTMLModElement>("ins", id, baseClass, job, scope) {
     fun cite(value: String) = attr("cite", value)
     fun cite(value: Flow<String>) = attr("cite", value)
 
@@ -762,8 +731,7 @@ open class Ins(id: String? = null, baseClass: String? = null, job: Job, scope: S
  * Exposes the JavaScript [HTMLModElement](https://developer.mozilla.org/en/docs/Web/API/HTMLModElement) to Kotlin
  */
 open class Del(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLModElement>("del", id, baseClass, job, scope),
-    WithText<HTMLModElement> {
+    HtmlTag<HTMLModElement>("del", id, baseClass, job, scope) {
     fun cite(value: String) = attr("cite", value)
     fun cite(value: Flow<String>) = attr("cite", value)
 
@@ -777,8 +745,7 @@ open class Del(id: String? = null, baseClass: String? = null, job: Job, scope: S
  * Exposes the JavaScript [HTMLOListElement](https://developer.mozilla.org/en/docs/Web/API/HTMLOListElement) to Kotlin
  */
 open class Ol(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLOListElement>("ol", id, baseClass, job, scope),
-    WithText<HTMLOListElement> {
+    HtmlTag<HTMLOListElement>("ol", id, baseClass, job, scope) {
     fun reversed(value: Boolean, trueValue: String = "") = attr("reversed", value, trueValue)
     fun reversed(value: Flow<Boolean>, trueValue: String = "") = attr("reversed", value, trueValue)
 
@@ -795,8 +762,7 @@ open class Ol(id: String? = null, baseClass: String? = null, job: Job, scope: Sc
  * Exposes the JavaScript [HTMLObjectElement](https://developer.mozilla.org/en/docs/Web/API/HTMLObjectElement) to Kotlin
  */
 open class Object(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLObjectElement>("object", id, baseClass, job, scope),
-    WithText<HTMLObjectElement> {
+    HtmlTag<HTMLObjectElement>("object", id, baseClass, job, scope) {
     fun data(value: String) = attr("data", value)
     fun data(value: Flow<String>) = attr("data", value)
 
@@ -825,8 +791,7 @@ open class Object(id: String? = null, baseClass: String? = null, job: Job, scope
  * Exposes the JavaScript [HTMLOptGroupElement](https://developer.mozilla.org/en/docs/Web/API/HTMLOptGroupElement) to Kotlin
  */
 open class Optgroup(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLOptGroupElement>("optgroup", id, baseClass, job, scope),
-    WithText<HTMLOptGroupElement> {
+    HtmlTag<HTMLOptGroupElement>("optgroup", id, baseClass, job, scope) {
     fun disabled(value: Boolean, trueValue: String = "") = attr("disabled", value, trueValue)
     fun disabled(value: Flow<Boolean>, trueValue: String = "") = attr("disabled", value, trueValue)
 
@@ -840,8 +805,7 @@ open class Optgroup(id: String? = null, baseClass: String? = null, job: Job, sco
  * Exposes the JavaScript [HTMLOptionElement](https://developer.mozilla.org/en/docs/Web/API/HTMLOptionElement) to Kotlin
  */
 open class Option(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLOptionElement>("option", id, baseClass, job, scope),
-    WithText<HTMLOptionElement> {
+    HtmlTag<HTMLOptionElement>("option", id, baseClass, job, scope) {
     fun disabled(value: Boolean, trueValue: String = "") = attr("disabled", value, trueValue)
     fun disabled(value: Flow<Boolean>, trueValue: String = "") = attr("disabled", value, trueValue)
 
@@ -871,8 +835,7 @@ open class Option(id: String? = null, baseClass: String? = null, job: Job, scope
  * Exposes the JavaScript [HTMLOutputElement](https://developer.mozilla.org/en/docs/Web/API/HTMLOutputElement) to Kotlin
  */
 open class Output(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLOutputElement>("output", id, baseClass, job, scope),
-    WithText<HTMLOutputElement> {
+    HtmlTag<HTMLOutputElement>("output", id, baseClass, job, scope) {
     fun name(value: String) = attr("name", value)
     fun name(value: Flow<String>) = attr("name", value)
 
@@ -895,16 +858,14 @@ open class Output(id: String? = null, baseClass: String? = null, job: Job, scope
  * Exposes the JavaScript [HTMLParagraphElement](https://developer.mozilla.org/en/docs/Web/API/HTMLParagraphElement) to Kotlin
  */
 open class P(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLParagraphElement>("p", id, baseClass, job, scope),
-    WithText<HTMLParagraphElement>
+    HtmlTag<HTMLParagraphElement>("p", id, baseClass, job, scope)
 
 
 /**
  * Exposes the JavaScript [HTMLParamElement](https://developer.mozilla.org/en/docs/Web/API/HTMLParamElement) to Kotlin
  */
 open class Param(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLParamElement>("param", id, baseClass, job, scope),
-    WithText<HTMLParamElement> {
+    HtmlTag<HTMLParamElement>("param", id, baseClass, job, scope) {
     fun name(value: String) = attr("name", value)
     fun name(value: Flow<String>) = attr("name", value)
 
@@ -917,24 +878,21 @@ open class Param(id: String? = null, baseClass: String? = null, job: Job, scope:
  * Exposes the JavaScript [HTMLPictureElement](https://developer.mozilla.org/en/docs/Web/API/HTMLPictureElement) to Kotlin
  */
 open class Picture(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLPictureElement>("picture", id, baseClass, job, scope),
-    WithText<HTMLPictureElement>
-
+    HtmlTag<HTMLPictureElement>("picture", id, baseClass, job, scope)
+    
 
 /**
  * Exposes the JavaScript [HTMLPreElement](https://developer.mozilla.org/en/docs/Web/API/HTMLPreElement) to Kotlin
  */
 open class Pre(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLPreElement>("pre", id, baseClass, job, scope),
-    WithText<HTMLPreElement>
-
+    HtmlTag<HTMLPreElement>("pre", id, baseClass, job, scope)
+    
 
 /**
  * Exposes the JavaScript [HTMLProgressElement](https://developer.mozilla.org/en/docs/Web/API/HTMLProgressElement) to Kotlin
  */
 open class Progress(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLProgressElement>("progress", id, baseClass, job, scope),
-    WithText<HTMLProgressElement> {
+    HtmlTag<HTMLProgressElement>("progress", id, baseClass, job, scope) {
     fun value(value: Double) = attr("value", value)
     fun value(value: Flow<Double>) = attr("value", value)
 
@@ -948,8 +906,7 @@ open class Progress(id: String? = null, baseClass: String? = null, job: Job, sco
  * Exposes the JavaScript [HTMLQuoteElement](https://developer.mozilla.org/en/docs/Web/API/HTMLQuoteElement) to Kotlin
  */
 open class Quote(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLQuoteElement>("quote", id, baseClass, job, scope),
-    WithText<HTMLQuoteElement> {
+    HtmlTag<HTMLQuoteElement>("quote", id, baseClass, job, scope) {
     fun cite(value: String) = attr("cite", value)
     fun cite(value: Flow<String>) = attr("cite", value)
 
@@ -960,8 +917,7 @@ open class Quote(id: String? = null, baseClass: String? = null, job: Job, scope:
  * Exposes the JavaScript [HTMLScriptElement](https://developer.mozilla.org/en/docs/Web/API/HTMLScriptElement) to Kotlin
  */
 open class Script(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLScriptElement>("script", id, baseClass, job, scope),
-    WithText<HTMLScriptElement> {
+    HtmlTag<HTMLScriptElement>("script", id, baseClass, job, scope) {
     fun src(value: String) = attr("src", value)
     fun src(value: Flow<String>) = attr("src", value)
 
@@ -994,7 +950,7 @@ open class Script(id: String? = null, baseClass: String? = null, job: Job, scope
  * Exposes the JavaScript [HTMLSelectElement](https://developer.mozilla.org/en/docs/Web/API/HTMLSelectElement) to Kotlin
  */
 open class Select(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLSelectElement>("select", id, baseClass, job, scope) {
+    HtmlTag<HTMLSelectElement>("select", id, baseClass, job, scope) {
     fun autocomplete(value: String) = attr("autocomplete", value)
     fun autocomplete(value: Flow<String>) = attr("autocomplete", value)
 
@@ -1031,24 +987,23 @@ open class Select(id: String? = null, baseClass: String? = null, job: Job, scope
  * Exposes the JavaScript [HTMLSpanElement](https://developer.mozilla.org/en/docs/Web/API/HTMLSpanElement) to Kotlin
  */
 open class Span(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLSpanElement>("span", id, baseClass, job, scope),
-    WithText<HTMLSpanElement>
+    HtmlTag<HTMLSpanElement>("span", id, baseClass, job, scope)
+    
 
 
 /**
  * Exposes the JavaScript [HTMLTableCaptionElement](https://developer.mozilla.org/en/docs/Web/API/HTMLTableCaptionElement) to Kotlin
  */
 open class Caption(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLTableCaptionElement>("caption", id, baseClass, job, scope),
-    WithText<HTMLTableCaptionElement>
+    HtmlTag<HTMLTableCaptionElement>("caption", id, baseClass, job, scope)
+    
 
 
 /**
  * Exposes the JavaScript [HTMLTableCellElement](https://developer.mozilla.org/en/docs/Web/API/HTMLTableCellElement) to Kotlin
  */
 open class Th(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLTableCellElement>("th", id, baseClass, job, scope),
-    WithText<HTMLTableCellElement> {
+    HtmlTag<HTMLTableCellElement>("th", id, baseClass, job, scope) {
     fun colSpan(value: Int) = attr("colSpan", value)
     fun colSpan(value: Flow<Int>) = attr("colSpan", value)
 
@@ -1071,8 +1026,7 @@ open class Th(id: String? = null, baseClass: String? = null, job: Job, scope: Sc
  * Exposes the JavaScript [HTMLTableCellElement](https://developer.mozilla.org/en/docs/Web/API/HTMLTableCellElement) to Kotlin
  */
 open class Td(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLTableCellElement>("td", id, baseClass, job, scope),
-    WithText<HTMLTableCellElement> {
+    HtmlTag<HTMLTableCellElement>("td", id, baseClass, job, scope) {
     fun colSpan(value: Int) = attr("colSpan", value)
     fun colSpan(value: Flow<Int>) = attr("colSpan", value)
 
@@ -1095,8 +1049,7 @@ open class Td(id: String? = null, baseClass: String? = null, job: Job, scope: Sc
  * Exposes the JavaScript [HTMLTableColElement](https://developer.mozilla.org/en/docs/Web/API/HTMLTableColElement) to Kotlin
  */
 open class Col(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLTableColElement>("col", id, baseClass, job, scope),
-    WithText<HTMLTableColElement> {
+    HtmlTag<HTMLTableColElement>("col", id, baseClass, job, scope) {
     fun span(value: Int) = attr("span", value)
     fun span(value: Flow<Int>) = attr("span", value)
 
@@ -1107,8 +1060,7 @@ open class Col(id: String? = null, baseClass: String? = null, job: Job, scope: S
  * Exposes the JavaScript [HTMLTableColElement](https://developer.mozilla.org/en/docs/Web/API/HTMLTableColElement) to Kotlin
  */
 open class Colgroup(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLTableColElement>("colgroup", id, baseClass, job, scope),
-    WithText<HTMLTableColElement> {
+    HtmlTag<HTMLTableColElement>("colgroup", id, baseClass, job, scope) {
     fun span(value: Int) = attr("span", value)
     fun span(value: Flow<Int>) = attr("span", value)
 
@@ -1119,44 +1071,43 @@ open class Colgroup(id: String? = null, baseClass: String? = null, job: Job, sco
  * Exposes the JavaScript [HTMLTableElement](https://developer.mozilla.org/en/docs/Web/API/HTMLTableElement) to Kotlin
  */
 open class Table(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLTableElement>("table", id, baseClass, job, scope)
+    HtmlTag<HTMLTableElement>("table", id, baseClass, job, scope)
 
 
 /**
  * Exposes the JavaScript [HTMLTableRowElement](https://developer.mozilla.org/en/docs/Web/API/HTMLTableRowElement) to Kotlin
  */
 open class Tr(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLTableRowElement>("tr", id, baseClass, job, scope),
-    WithText<HTMLTableRowElement>
+    HtmlTag<HTMLTableRowElement>("tr", id, baseClass, job, scope)
+    
 
 
 /**
  * Exposes the JavaScript [HTMLTableSectionElement](https://developer.mozilla.org/en/docs/Web/API/HTMLTableSectionElement) to Kotlin
  */
 open class TFoot(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLTableSectionElement>("tfoot", id, baseClass, job, scope)
+    HtmlTag<HTMLTableSectionElement>("tfoot", id, baseClass, job, scope)
 
 
 /**
  * Exposes the JavaScript [HTMLTableSectionElement](https://developer.mozilla.org/en/docs/Web/API/HTMLTableSectionElement) to Kotlin
  */
 open class THead(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLTableSectionElement>("thead", id, baseClass, job, scope)
+    HtmlTag<HTMLTableSectionElement>("thead", id, baseClass, job, scope)
 
 
 /**
  * Exposes the JavaScript [HTMLTableSectionElement](https://developer.mozilla.org/en/docs/Web/API/HTMLTableSectionElement) to Kotlin
  */
 open class TBody(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLTableSectionElement>("tbody", id, baseClass, job, scope)
+    HtmlTag<HTMLTableSectionElement>("tbody", id, baseClass, job, scope)
 
 
 /**
  * Exposes the JavaScript [HTMLTextAreaElement](https://developer.mozilla.org/en/docs/Web/API/HTMLTextAreaElement) to Kotlin
  */
 open class TextArea(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLTextAreaElement>("textarea", id, baseClass, job, scope),
-    WithText<HTMLTextAreaElement> {
+    HtmlTag<HTMLTextAreaElement>("textarea", id, baseClass, job, scope) {
     fun autocomplete(value: String) = attr("autocomplete", value)
     fun autocomplete(value: Flow<String>) = attr("autocomplete", value)
 
@@ -1218,8 +1169,7 @@ open class TextArea(id: String? = null, baseClass: String? = null, job: Job, sco
  * Exposes the JavaScript [HTMLTimeElement](https://developer.mozilla.org/en/docs/Web/API/HTMLTimeElement) to Kotlin
  */
 open class Time(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLTimeElement>("time", id, baseClass, job, scope),
-    WithText<HTMLTimeElement> {
+    HtmlTag<HTMLTimeElement>("time", id, baseClass, job, scope) {
     fun dateTime(value: String) = attr("dateTime", value)
     fun dateTime(value: Flow<String>) = attr("dateTime", value)
 
@@ -1230,8 +1180,7 @@ open class Time(id: String? = null, baseClass: String? = null, job: Job, scope: 
  * Exposes the JavaScript [HTMLTrackElement](https://developer.mozilla.org/en/docs/Web/API/HTMLTrackElement) to Kotlin
  */
 open class Track(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLTrackElement>("track", id, baseClass, job, scope),
-    WithText<HTMLTrackElement> {
+    HtmlTag<HTMLTrackElement>("track", id, baseClass, job, scope) {
     fun kind(value: String) = attr("kind", value)
     fun kind(value: Flow<String>) = attr("kind", value)
 
@@ -1254,14 +1203,14 @@ open class Track(id: String? = null, baseClass: String? = null, job: Job, scope:
  * Exposes the JavaScript [HTMLUListElement](https://developer.mozilla.org/en/docs/Web/API/HTMLUListElement) to Kotlin
  */
 open class Ul(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLUListElement>("ul", id, baseClass, job, scope)
+    HtmlTag<HTMLUListElement>("ul", id, baseClass, job, scope)
 
 
 /**
  * Exposes the JavaScript [SVGElement](https://developer.mozilla.org/en-US/docs/Web/API/SVGElement) to Kotlin
  */
 class Svg(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<SVGElement>("", id, baseClass, job, scope) {
+    HtmlTag<SVGElement>("", id, baseClass, job, scope) {
 
     override fun createDomNode(): SVGElement =
         document.createElementNS(SVG_XMLNS, "svg").unsafeCast<SVGElement>().apply {
@@ -1292,7 +1241,7 @@ const val SVG_XMLNS = "http://www.w3.org/2000/svg"
  * Exposes the JavaScript Path(https://developer.mozilla.org/en-US/docs/Web/SVG/Tutorial/Paths) to Kotlin
  */
 class Path(id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<SVGPathElement>("", id, baseClass, job, scope) {
+    HtmlTag<SVGPathElement>("", id, baseClass, job, scope) {
 
     override fun createDomNode(): SVGPathElement =
         document.createElementNS(SVG_XMLNS, "path").unsafeCast<SVGPathElement>().apply {
@@ -1311,948 +1260,8 @@ class Path(id: String? = null, baseClass: String? = null, job: Job, scope: Scope
 }
 
 /**
- * Special [Tag] for HTML5 with no attributes
+ * Special [HtmlTag] for HTML5 with no attributes
  */
 open class TextElement(tagName: String, id: String? = null, baseClass: String? = null, job: Job, scope: Scope) :
-    Tag<HTMLElement>(tagName, id, baseClass, job, scope), WithText<HTMLElement>
+    HtmlTag<HTMLElement>(tagName, id, baseClass, job, scope)
 
-
-/**
- * Context for rendering static and dynamical content
- */
-interface RenderContext : WithJob, WithScope {
-
-    /**
-     * Renders the data of a [Flow] as [Tag]s to the DOM.
-     *
-     * @receiver [Flow] containing the data
-     * @param into target to mount content to. If not set a child div is added to the [Tag] this method is called on
-     * @param content [RenderContext] for rendering the data to the DOM
-     */
-    fun <V> Flow<V>.render(into: Tag<HTMLElement>? = null, content: RenderContext.(V) -> Unit) {
-        val target = into?.apply(SET_MOUNT_POINT_DATA_ATTRIBUTE)
-            ?: div(MOUNT_POINT_STYLE_CLASS, content = SET_MOUNT_POINT_DATA_ATTRIBUTE)
-
-        val mountContext = MountContext(Job(job), target)
-
-        mountSimple(job, this) {
-            mountContext.job.cancelChildren()
-            mountContext.runBeforeUnmounts()
-            target.domNode.clear()
-            content(mountContext, it)
-            mountContext.runAfterMounts()
-        }
-    }
-
-    /**
-     * Renders each element of a [Flow]s content.
-     * Internally the [Patch]es are determined using Myer's diff-algorithm.
-     * This allows the detection of moves. Keep in mind, that no [Patch] is derived,
-     * when an element stays the same, but changes its internal values.
-     *
-     * @param idProvider function to identify a unique entity in the list
-     * @param into target to mount content to. If not set a child div is added to the [Tag] this method is called on
-     * @param content [RenderContext] for rendering the data to the DOM
-     */
-    fun <V> Flow<List<V>>.renderEach(
-        idProvider: IdProvider<V, *>? = null,
-        into: Tag<HTMLElement>? = null,
-        content: RenderContext.(V) -> Tag<HTMLElement>
-    ) {
-        mountPatches(into, this) { upstreamValues, mountPoints ->
-            upstreamValues.scan(Pair(emptyList(), emptyList())) { acc: Pair<List<V>, List<V>>, new ->
-                Pair(acc.second, new)
-            }.map { (old, new) ->
-                val diff = if (idProvider != null) Myer.diff(old, new, idProvider) else Myer.diff(old, new)
-                diff.map { patch ->
-                    patch.map(job) { value, newJob ->
-                        val mountPoint = BuildContext(newJob, scope)
-                        content(mountPoint, value).also {
-                            mountPoints[it.domNode] = mountPoint
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * Renders each element of a [Store]s [List] content.
-     * Internally the [Patch]es are determined using Myer's diff-algorithm.
-     * This allows the detection of moves. Keep in mind, that no [Patch] is derived,
-     * when an element stays the same, but changes its internal values.
-     *
-     * @param idProvider function to identify a unique entity in the list
-     * @param into target to mount content to. If not set a child div is added to the [Tag] this method is called on
-     * @param content [RenderContext] for rendering the data to the DOM
-     */
-    fun <V> Store<List<V>>.renderEach(
-        idProvider: IdProvider<V, *>,
-        into: Tag<HTMLElement>? = null,
-        content: RenderContext.(Store<V>) -> Tag<HTMLElement>
-    ) {
-        data.renderEach(idProvider, into) { value ->
-            content(this@renderEach.sub(value, idProvider))
-        }
-    }
-
-    /**
-     * Converts the content of a [Flow] to [String] by using [toString] method.
-     *
-     * @receiver [Flow] with content
-     * @return [Flow] with content as [String]
-     */
-    fun <T> Flow<T>.asString(): Flow<String> = this.map { it.toString() }
-
-    fun <N : Node, W : WithDomNode<N>> register(element: W, content: (W) -> Unit): W
-
-    /**
-     * Evaluates the scope context and initializes a [ScopeContext]
-     * for setting new entries to the scope.
-     *
-     * @param context to evaluate
-     */
-    private inline fun evalScope(context: (ScopeContext.() -> Unit)): Scope {
-        return ScopeContext(this@RenderContext.scope).apply(context).scope
-    }
-
-    /**
-     * Creates a custom [Tag] with the provided [content].
-     *
-     * @param tagName Name of the [Tag] in DOM
-     * @param content content scope for inner [Tag]s
-     * @return custom [Tag]
-     */
-    fun custom(
-        tagName: String,
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Tag<*>.() -> Unit
-    ): Tag<*> =
-        register(Tag(tagName, id, baseClass, job, evalScope(scope)), content)
-
-    fun a(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: A.() -> Unit
-    ): A =
-        register(A(id, baseClass, job, evalScope(scope)), content)
-
-    fun area(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Area.() -> Unit
-    ): Area =
-        register(Area(id, baseClass, job, evalScope(scope)), content)
-
-    fun br(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Br.() -> Unit
-    ): Br =
-        register(Br(id, baseClass, job, evalScope(scope)), content)
-
-    fun button(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Button.() -> Unit
-    ): Button =
-        register(Button(id, baseClass, job, evalScope(scope)), content)
-
-    fun canvas(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Canvas.() -> Unit
-    ): Canvas =
-        register(Canvas(id, baseClass, job, evalScope(scope)), content)
-
-    fun dl(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Dl.() -> Unit
-    ): Dl =
-        register(Dl(id, baseClass, job, evalScope(scope)), content)
-
-    fun dt(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("dt", id, baseClass, job, evalScope(scope)), content)
-
-    fun dd(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("dd", id, baseClass, job, evalScope(scope)), content)
-
-    fun data(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Data.() -> Unit
-    ): Data =
-        register(Data(id, baseClass, job, evalScope(scope)), content)
-
-    fun datalist(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: DataList.() -> Unit
-    ): DataList =
-        register(DataList(id, baseClass, job, evalScope(scope)), content)
-
-    fun details(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Details.() -> Unit
-    ): Details =
-        register(Details(id, baseClass, job, evalScope(scope)), content)
-
-    fun dialog(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Dialog.() -> Unit
-    ): Dialog =
-        register(Dialog(id, baseClass, job, evalScope(scope)), content)
-
-    fun div(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Div.() -> Unit
-    ): Div =
-        register(Div(id, baseClass, job, evalScope(scope)), content)
-
-    fun embed(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Embed.() -> Unit
-    ): Embed =
-        register(Embed(id, baseClass, job, evalScope(scope)), content)
-
-    fun fieldset(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: FieldSet.() -> Unit
-    ): FieldSet =
-        register(FieldSet(id, baseClass, job, evalScope(scope)), content)
-
-    fun form(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Form.() -> Unit
-    ): Form =
-        register(Form(id, baseClass, job, evalScope(scope)), content)
-
-    fun hr(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Hr.() -> Unit
-    ): Hr =
-        register(Hr(id, baseClass, job, evalScope(scope)), content)
-
-    fun h1(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: H.() -> Unit
-    ): H =
-        register(H(1, id, baseClass, job, evalScope(scope)), content)
-
-    fun h2(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: H.() -> Unit
-    ): H =
-        register(H(2, id, baseClass, job, evalScope(scope)), content)
-
-    fun h3(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: H.() -> Unit
-    ): H =
-        register(H(3, id, baseClass, job, evalScope(scope)), content)
-
-    fun h4(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: H.() -> Unit
-    ): H =
-        register(H(4, id, baseClass, job, evalScope(scope)), content)
-
-    fun h5(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: H.() -> Unit
-    ): H =
-        register(H(5, id, baseClass, job, evalScope(scope)), content)
-
-    fun h6(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: H.() -> Unit
-    ): H =
-        register(H(6, id, baseClass, job, evalScope(scope)), content)
-
-    fun iframe(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: IFrame.() -> Unit
-    ): IFrame =
-        register(IFrame(id, baseClass, job, evalScope(scope)), content)
-
-    fun img(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Img.() -> Unit
-    ): Img =
-        register(Img(id, baseClass, job, evalScope(scope)), content)
-
-    fun input(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Input.() -> Unit
-    ): Input =
-        register(Input(id, baseClass, job, evalScope(scope)), content)
-
-    fun li(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Li.() -> Unit
-    ): Li =
-        register(Li(id, baseClass, job, evalScope(scope)), content)
-
-    fun label(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Label.() -> Unit
-    ): Label =
-        register(Label(id, baseClass, job, evalScope(scope)), content)
-
-    fun legend(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Legend.() -> Unit
-    ): Legend =
-        register(Legend(id, baseClass, job, evalScope(scope)), content)
-
-    fun map(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Map.() -> Unit
-    ): Map =
-        register(Map(id, baseClass, job, evalScope(scope)), content)
-
-    fun audio(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Audio.() -> Unit
-    ): Audio =
-        register(Audio(id, baseClass, job, evalScope(scope)), content)
-
-    fun video(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Video.() -> Unit
-    ): Video =
-        register(Video(id, baseClass, job, evalScope(scope)), content)
-
-    fun meter(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Meter.() -> Unit
-    ): Meter =
-        register(Meter(id, baseClass, job, evalScope(scope)), content)
-
-    fun ins(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Ins.() -> Unit
-    ): Ins =
-        register(Ins(id, baseClass, job, evalScope(scope)), content)
-
-    fun del(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Del.() -> Unit
-    ): Del =
-        register(Del(id, baseClass, job, evalScope(scope)), content)
-
-    fun ol(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Ol.() -> Unit
-    ): Ol =
-        register(Ol(id, baseClass, job, evalScope(scope)), content)
-
-    fun `object`(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Object.() -> Unit
-    ): Object =
-        register(Object(id, baseClass, job, evalScope(scope)), content)
-
-    fun optgroup(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Optgroup.() -> Unit
-    ): Optgroup =
-        register(Optgroup(id, baseClass, job, evalScope(scope)), content)
-
-    fun option(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Option.() -> Unit
-    ): Option =
-        register(Option(id, baseClass, job, evalScope(scope)), content)
-
-    fun output(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Output.() -> Unit
-    ): Output =
-        register(Output(id, baseClass, job, evalScope(scope)), content)
-
-    fun p(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: P.() -> Unit
-    ): P =
-        register(P(id, baseClass, job, evalScope(scope)), content)
-
-    fun param(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Param.() -> Unit
-    ): Param =
-        register(Param(id, baseClass, job, evalScope(scope)), content)
-
-    fun picture(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Picture.() -> Unit
-    ): Picture =
-        register(Picture(id, baseClass, job, evalScope(scope)), content)
-
-    fun pre(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Pre.() -> Unit
-    ): Pre =
-        register(Pre(id, baseClass, job, evalScope(scope)), content)
-
-    fun progress(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Progress.() -> Unit
-    ): Progress =
-        register(Progress(id, baseClass, job, evalScope(scope)), content)
-
-    fun quote(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Quote.() -> Unit
-    ): Quote =
-        register(Quote(id, baseClass, job, evalScope(scope)), content)
-
-    fun script(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Script.() -> Unit
-    ): Script =
-        register(Script(id, baseClass, job, evalScope(scope)), content)
-
-    fun select(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Select.() -> Unit
-    ): Select =
-        register(Select(id, baseClass, job, evalScope(scope)), content)
-
-    fun span(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Span.() -> Unit
-    ): Span =
-        register(Span(id, baseClass, job, evalScope(scope)), content)
-
-    fun caption(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Caption.() -> Unit
-    ): Caption =
-        register(Caption(id, baseClass, job, evalScope(scope)), content)
-
-    fun th(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Th.() -> Unit
-    ): Th =
-        register(Th(id, baseClass, job, evalScope(scope)), content)
-
-    fun td(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Td.() -> Unit
-    ): Td =
-        register(Td(id, baseClass, job, evalScope(scope)), content)
-
-    fun col(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Col.() -> Unit
-    ): Col =
-        register(Col(id, baseClass, job, evalScope(scope)), content)
-
-    fun colgroup(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Colgroup.() -> Unit
-    ): Colgroup =
-        register(Colgroup(id, baseClass, job, evalScope(scope)), content)
-
-    fun table(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Table.() -> Unit
-    ): Table =
-        register(Table(id, baseClass, job, evalScope(scope)), content)
-
-    fun tr(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Tr.() -> Unit
-    ): Tr =
-        register(Tr(id, baseClass, job, evalScope(scope)), content)
-
-    fun tfoot(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TFoot.() -> Unit
-    ): TFoot =
-        register(TFoot(id, baseClass, job, evalScope(scope)), content)
-
-    fun thead(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: THead.() -> Unit
-    ): THead =
-        register(THead(id, baseClass, job, evalScope(scope)), content)
-
-    fun tbody(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TBody.() -> Unit
-    ): TBody =
-        register(TBody(id, baseClass, job, evalScope(scope)), content)
-
-    fun textarea(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextArea.() -> Unit
-    ): TextArea =
-        register(TextArea(id, baseClass, job, evalScope(scope)), content)
-
-    fun time(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Time.() -> Unit
-    ): Time =
-        register(Time(id, baseClass, job, evalScope(scope)), content)
-
-    fun track(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Track.() -> Unit
-    ): Track =
-        register(Track(id, baseClass, job, evalScope(scope)), content)
-
-    fun ul(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Ul.() -> Unit
-    ): Ul =
-        register(Ul(id, baseClass, job, evalScope(scope)), content)
-
-    fun address(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("address", id, baseClass, job, evalScope(scope)), content)
-
-    fun article(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("article", id, baseClass, job, evalScope(scope)), content)
-
-    fun aside(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("aside", id, baseClass, job, evalScope(scope)), content)
-
-    fun bdi(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("bdi", id, baseClass, job, evalScope(scope)), content)
-
-    fun figcaption(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("figcaption", id, baseClass, job, evalScope(scope)), content)
-
-    fun figure(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("figure", id, baseClass, job, evalScope(scope)), content)
-
-    fun footer(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("footer", id, baseClass, job, evalScope(scope)), content)
-
-    fun header(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("header", id, baseClass, job, evalScope(scope)), content)
-
-    fun main(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("main", id, baseClass, job, evalScope(scope)), content)
-
-    fun mark(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("mark", id, baseClass, job, evalScope(scope)), content)
-
-    fun nav(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("nav", id, baseClass, job, evalScope(scope)), content)
-
-    fun noscript(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("noscript", id, baseClass, job, evalScope(scope)), content)
-
-    fun rp(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("rp", id, baseClass, job, evalScope(scope)), content)
-
-    fun rt(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("rt", id, baseClass, job, evalScope(scope)), content)
-
-    fun ruby(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("ruby", id, baseClass, job, evalScope(scope)), content)
-
-    fun section(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("section", id, baseClass, job, evalScope(scope)), content)
-
-    fun summary(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("summary", id, baseClass, job, evalScope(scope)), content)
-
-    fun wbr(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("wbr", id, baseClass, job, evalScope(scope)), content)
-
-    fun blockquote(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("blockquote", id, baseClass, job, evalScope(scope)), content)
-
-    fun em(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("em", id, baseClass, job, evalScope(scope)), content)
-
-    fun strong(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("strong", id, baseClass, job, evalScope(scope)), content)
-
-    fun small(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("small", id, baseClass, job, evalScope(scope)), content)
-
-    fun s(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("s", id, baseClass, job, evalScope(scope)), content)
-
-    fun cite(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("cite", id, baseClass, job, evalScope(scope)), content)
-
-    fun q(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("q", id, baseClass, job, evalScope(scope)), content)
-
-    fun dfn(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("dfn", id, baseClass, job, evalScope(scope)), content)
-
-    fun abbr(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("abbr", id, baseClass, job, evalScope(scope)), content)
-
-    fun code(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("code", id, baseClass, job, evalScope(scope)), content)
-
-    fun `var`(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("var", id, baseClass, job, evalScope(scope)), content)
-
-    fun samp(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("samp", id, baseClass, job, evalScope(scope)), content)
-
-    fun kbd(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("kbd", id, baseClass, job, evalScope(scope)), content)
-
-    fun sub(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("sub", id, baseClass, job, evalScope(scope)), content)
-
-    fun sup(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("sup", id, baseClass, job, evalScope(scope)), content)
-
-    fun i(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("i", id, baseClass, job, evalScope(scope)), content)
-
-    fun b(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("b", id, baseClass, job, evalScope(scope)), content)
-
-    fun u(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("u", id, baseClass, job, evalScope(scope)), content)
-
-    fun bdo(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("bdo", id, baseClass, job, evalScope(scope)), content)
-
-    fun command(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: TextElement.() -> Unit
-    ): TextElement =
-        register(TextElement("command", id, baseClass, job, evalScope(scope)), content)
-
-    fun svg(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Svg.() -> Unit
-    ): Svg =
-        register(Svg(id, baseClass, job = job, evalScope(scope)), content)
-
-    fun path(
-        baseClass: String? = null,
-        id: String? = null,
-        scope: (ScopeContext.() -> Unit) = {},
-        content: Path.() -> Unit
-    ): Path =
-        register(Path(id, baseClass, job = job, evalScope(scope)), content)
-}
