+++ core/src/jsTest/kotlin/dev/fritz2/webcomponents/webcomponent.kt
@@ -1,7 +1,7 @@
 package dev.fritz2.webcomponents
 
 import dev.fritz2.binding.storeOf
-import dev.fritz2.dom.Tag
+import dev.fritz2.dom.HtmlTag
 import dev.fritz2.dom.html.RenderContext
 import dev.fritz2.test.initDocument
 import dev.fritz2.test.runTest
@@ -42,7 +42,7 @@ class WebComponentTests {
     }
 
     object AttributeComponent : WebComponent<HTMLElement>() {
-        override fun RenderContext.init(element: HTMLElement, shadowRoot: ShadowRoot): Tag<HTMLElement> {
+        override fun RenderContext.init(element: HTMLElement, shadowRoot: ShadowRoot): HtmlTag<HTMLElement> {
             val store = storeOf("Initial")
             attributeChanges("test") handledBy { store.update(it) }
