+++ headless/src/jsMain/kotlin/dev/fritz2/headless/components/modal.kt
@@ -1,9 +1,9 @@
 package dev.fritz2.headless.components
 
-import dev.fritz2.dom.Tag
+import dev.fritz2.dom.HtmlTag
 import dev.fritz2.dom.html.*
-import dev.fritz2.identification.Id
 import dev.fritz2.headless.foundation.*
+import dev.fritz2.identification.Id
 import org.w3c.dom.HTMLElement
 
 class HeadlessModal(val renderContext: RenderContext) : RenderContext by renderContext,
@@ -12,7 +12,7 @@ class HeadlessModal(val renderContext: RenderContext) : RenderContext by renderC
     var restoreFocus: Boolean = true
     var setInitialFocus: Boolean = true
 
-    private var panel: (RenderContext.() -> Tag<HTMLElement>)? = null
+    private var panel: (RenderContext.() -> HtmlTag<HTMLElement>)? = null
 
     fun render() {
         opened.render {
@@ -24,14 +24,14 @@ class HeadlessModal(val renderContext: RenderContext) : RenderContext by renderC
         }
     }
 
-    inner class ModalPanel<C : Tag<HTMLElement>>(
+    inner class ModalPanel<C : HtmlTag<HTMLElement>>(
         val tag: C,
         private val id: String? = null
     ) : RenderContext by tag {
         val componentId: String by lazy { id ?: Id.next() }
 
-        private var title: Tag<HTMLElement>? = null
-        private var description: Tag<HTMLElement>? = null
+        private var title: HtmlTag<HTMLElement>? = null
+        private var description: HtmlTag<HTMLElement>? = null
 
         fun C.render() {
             attr("id", componentId)
@@ -41,7 +41,7 @@ class HeadlessModal(val renderContext: RenderContext) : RenderContext by renderC
             description?.let { attr(Aria.describedby, it.id) }
         }
 
-        fun <CO : Tag<HTMLElement>> RenderContext.modalOverlay(
+        fun <CO : HtmlTag<HTMLElement>> RenderContext.modalOverlay(
             classes: String? = null,
             scope: (ScopeContext.() -> Unit) = {},
             tag: TagFactory<CO>,
@@ -57,7 +57,7 @@ class HeadlessModal(val renderContext: RenderContext) : RenderContext by renderC
             content: Div.() -> Unit
         ) = modalOverlay(classes, scope, RenderContext::div, content)
 
-        fun <CT : Tag<HTMLElement>> RenderContext.modalTitle(
+        fun <CT : HtmlTag<HTMLElement>> RenderContext.modalTitle(
             classes: String? = null,
             scope: (ScopeContext.() -> Unit) = {},
             tag: TagFactory<CT>,
@@ -70,7 +70,7 @@ class HeadlessModal(val renderContext: RenderContext) : RenderContext by renderC
             content: H.() -> Unit
         ) = modalTitle(classes, scope, RenderContext::h2, content)
 
-        fun <CD : Tag<HTMLElement>> RenderContext.modalDescription(
+        fun <CD : HtmlTag<HTMLElement>> RenderContext.modalDescription(
             classes: String? = null,
             scope: (ScopeContext.() -> Unit) = {},
             tag: TagFactory<CD>,
@@ -84,7 +84,7 @@ class HeadlessModal(val renderContext: RenderContext) : RenderContext by renderC
         ) = modalDescription(classes, scope, RenderContext::p, content)
     }
 
-    fun <C : Tag<HTMLElement>> RenderContext.modalPanel(
+    fun <C : HtmlTag<HTMLElement>> RenderContext.modalPanel(
         classes: String? = null,
         id: String? = null,
         internalScope: (ScopeContext.() -> Unit) = {},
