+++ headless/src/jsMain/kotlin/dev/fritz2/headless/components/tooltip.kt
@@ -1,7 +1,7 @@
 package dev.fritz2.headless.components
 
 import dev.fritz2.binding.storeOf
-import dev.fritz2.dom.Tag
+import dev.fritz2.dom.HtmlTag
 import dev.fritz2.dom.Window
 import dev.fritz2.dom.html.*
 import dev.fritz2.dom.merge
@@ -17,8 +17,8 @@ import org.w3c.dom.HTMLElement
 
 
 @Suppress("EXPERIMENTAL_IS_NOT_ENABLED")
-class HeadlessTooltip<C : Tag<HTMLElement>>(
-    val target: Tag<HTMLElement>,
+class HeadlessTooltip<C : HtmlTag<HTMLElement>>(
+    val target: HtmlTag<HTMLElement>,
     val tag: C,
 ) : OpenClose by OpenCloseDelegate() {
 
@@ -70,7 +70,7 @@ class HeadlessTooltip<C : Tag<HTMLElement>>(
     }
 }
 
-fun <C : Tag<HTMLElement>> Tag<HTMLElement>.headlessTooltip(
+fun <C : HtmlTag<HTMLElement>> HtmlTag<HTMLElement>.headlessTooltip(
     classes: String? = null,
     id: String? = null,
     scope: (ScopeContext.() -> Unit) = {},
@@ -86,7 +86,7 @@ fun <C : Tag<HTMLElement>> Tag<HTMLElement>.headlessTooltip(
 }
 
 
-fun Tag<HTMLElement>.headlessTooltip(
+fun HtmlTag<HTMLElement>.headlessTooltip(
     classes: String? = null,
     id: String? = null,
     internalScope: (ScopeContext.() -> Unit) = {},
