+++ headless/src/jsMain/kotlin/dev/fritz2/headless/foundation/openclose.kt
@@ -1,12 +1,11 @@
 package dev.fritz2.headless.foundation
 
 import dev.fritz2.binding.SimpleHandler
-import dev.fritz2.dom.Tag
+import dev.fritz2.dom.HtmlTag
 import dev.fritz2.dom.html.Keys
 import dev.fritz2.dom.html.shortcutOf
 
 import dev.fritz2.headless.hooks.DatabindingHook
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.*
 import org.w3c.dom.HTMLElement
 
@@ -18,8 +17,8 @@ interface OpenClose {
     val toggle: SimpleHandler<Unit>
 }
 
-class OpenCloseDatabindingHook : DatabindingHook<Tag<HTMLElement>, Unit, Boolean>() {
-    override fun Tag<HTMLElement>.render(payload: Unit) {
+class OpenCloseDatabindingHook : DatabindingHook<HtmlTag<HTMLElement>, Unit, Boolean>() {
+    override fun HtmlTag<HTMLElement>.render(payload: Unit) {
         handler?.invoke(data.flatMapLatest { state ->
             merge(
                 clicks.events,
