+++ headless/src/jsMain/kotlin/dev/fritz2/headless/hooks/databinding.kt
@@ -1,20 +1,17 @@
 package dev.fritz2.headless.hooks
 
-import dev.fritz2.binding.RootStore
 import dev.fritz2.binding.Store
-import dev.fritz2.dom.Tag
-import dev.fritz2.dom.html.Keys
-import dev.fritz2.dom.html.Shortcut
+import dev.fritz2.dom.HtmlTag
 import dev.fritz2.dom.html.handledBy
-import dev.fritz2.dom.html.shortcutOf
 import dev.fritz2.headless.validation.ComponentValidationMessage
 import dev.fritz2.headless.validation.Severity
 import dev.fritz2.headless.validation.validationMessages
-import kotlinx.coroutines.flow.*
-import org.w3c.dom.HTMLElement
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flowOf
+import kotlinx.coroutines.flow.map
 
 
-abstract class DatabindingHook<C : Tag<*>, P, T> : BasicHook<C, Unit, P>(), Usable<DatabindingHook<C, P, T>> {
+abstract class DatabindingHook<C : HtmlTag<*>, P, T> : BasicHook<C, Unit, P>(), Usable<DatabindingHook<C, P, T>> {
     lateinit var data: Flow<T>
     var id: String? = null
     var handler: ((Flow<T>) -> Unit)? = null
@@ -65,6 +62,6 @@ abstract class DatabindingHook<C : Tag<*>, P, T> : BasicHook<C, Unit, P>(), Usab
     }
 }
 
-abstract class ItemDatabindingHook<C : Tag<*>, P, T> : DatabindingHook<C, P, T>() {
+abstract class ItemDatabindingHook<C : HtmlTag<*>, P, T> : DatabindingHook<C, P, T>() {
     abstract fun isSelected(item: P): Flow<Boolean>
 }
